Event show page
We want to be able to market to specific events in online ads.

This changeset introduces a new events show page at /events/:id where
:id is the last url part of the eventbrite url. For example, if the
event name is Turing Try Coding, we'll get an eventbrite url ending in
something like 'turing-try-coding-124445'. We use that string as our :id
param and then pluck off the integers at the end of it (after the last
dash) to look up the event in our internal cache of the event data.

https://trello.com/c/eCA7eRyA/124-create-event-specific-page-in-tryturing